requirements.txt
requests
mypy
black
isort
pylint
pre-commit
types-requests
PyYAML